+++ lib/flight.py
@@ -20,12 +20,15 @@ class Flight:
     The flight time is automatically translated from the flight's local timezone to UTC.
     """
 
-    def __init__(self, flight_info: Dict[str, Any], confirmation_number: str) -> None:
+    def __init__(self, flight_info: JSON, reservation_info: JSON, confirmation_number: str) -> None:
         self.confirmation_number = confirmation_number
         self.departure_airport = flight_info["departureAirport"]["name"]
         self.destination_airport = flight_info["arrivalAirport"]["name"]
         self.flight_number = self._get_flight_number(flight_info["flights"])
         self.is_same_day = False
+
+        # Cached for use by the fare checker
+        self.reservation_info = reservation_info
 
         # Track to notify the user of filling out their passport information.
         # Southwest only fills the country's value for international flights