+++ tests/integration/test_check_in.py
@@ -24,7 +24,7 @@ def handler(mocker: MockerFixture) -> None:
         "departureTime": "14:40",
         "flights": [{"number": "100"}],
     }
-    flight = Flight(flight_info, "TEST")
+    flight = Flight(flight_info, {}, "TEST")
     # Make sure it isn't affected by local time
     flight.departure_time = datetime(2021, 12, 6, 14, 40)
     return CheckInHandler(mock_scheduler, flight, Lock())